Fix: Hide methodology selection page, delete data on methodology change
