+++ app/src/backend/ActivityService.ts
@@ -406,9 +406,11 @@ export default class ActivityService {
   public static async deleteAllActivitiesInSubsector({
     subsectorId,
     inventoryId,
+    referenceNumber,
   }: {
-    subsectorId: string;
+    subsectorId?: string;
     inventoryId: string;
+    referenceNumber?: string;
   }): Promise<number> {
     const inventoryValues = await db.models.InventoryValue.findAll({
       where: {
@@ -427,8 +429,10 @@ export default class ActivityService {
 
     return await db.models.InventoryValue.destroy({
       where: {
-        subSectorId: subsectorId,
         inventoryId,
+        ...(referenceNumber
+          ? { gpcReferenceNumber: referenceNumber }
+          : { subSectorId: subsectorId }),
       },
     });
   }
