+++ app/src/components/Tabs/Activity/activity-tab.tsx
@@ -1,49 +1,18 @@
-import {
-  Box,
-  Button,
-  Card,
-  Icon,
-  IconButton,
-  Link,
-  Popover,
-  PopoverArrow,
-  PopoverBody,
-  PopoverContent,
-  PopoverTrigger,
-  SimpleGrid,
-  Switch,
-  TabPanel,
-  Text,
-  useDisclosure,
-} from "@chakra-ui/react";
+import { Box, Switch, TabPanel, Text } from "@chakra-ui/react";
 import React, { FC, useMemo, useState } from "react";
 import HeadingText from "../../heading-text";
-import { AddIcon } from "@chakra-ui/icons";
-import { MdMoreVert } from "react-icons/md";
-import SuggestedActivityCard from "../../Cards/suggested-activities-card";
 import { TFunction } from "i18next";
-import { FiTrash2 } from "react-icons/fi";
-import { FaNetworkWired } from "react-icons/fa";
-import { Trans } from "react-i18next";
-import ActivityFormModal from "../../Modals/activity-modal/activity-form-modal";
-import ChangeMethodology from "../../Modals/change-methodology";
-import DeleteAllActivitiesModal from "../../Modals/delete-all-activities-modal";
-import { api } from "@/services/api";
-import ActivityAccordion from "./activity-accordion";
 import ScopeUnavailable from "./scope-unavailable";
 import {
+  DirectMeasure,
   MANUAL_INPUT_HIERARCHY,
   Methodology,
   SuggestedActivity,
 } from "@/util/form-schema";
-import MethodologyCard from "@/components/Cards/methodology-card";
-import { ActivityData } from "@/models/ActivityData";
 import { ActivityValue } from "@/models/ActivityValue";
-import { DataConnectIcon } from "@/components/icons";
-import { convertKgToTonnes, getInputMethodology } from "@/util/helpers";
-import DirectMeasureTable from "./direct-measure-table";
-import DeleteActivityModal from "@/components/Modals/delete-activity-modal";
 import { InventoryValue } from "@/models/InventoryValue";
+import EmissionDataSection from "@/components/Tabs/Activity/emission-data-section";
+import SelectMethodology from "@/components/Tabs/Activity/select-methodology";
 
 interface ActivityTabProps {
   t: TFunction;
@@ -82,23 +51,54 @@ const ActivityTab: FC<ActivityTabProps> = ({
     totalEmissions += parseInt(activity?.co2eq);
   });
 
-  const [selectedActivity, setSelectedActivity] = useState<
-    SuggestedActivity | undefined
-  >();
-  const [isMethodologySelected, setIsMethodologySelected] = useState(false);
+  const [isMethodologySelected, setIsMethodologySelected] = useState();
   const [selectedMethodology, setSelectedMethodology] = useState("");
   const [isUnavailableChecked, setIsChecked] = useState<boolean>(false);
-  const [hasActivityData, setHasActivityData] = useState<boolean>(false);
-  const [methodology, setMethodology] = useState<Methodology>();
-  const [selectedActivityValue, setSelectedActivityValue] =
-    useState<ActivityValue>();
 
   const refNumberWithScope = referenceNumber + "." + (filteredScope || 1);
 
-  const getfilteredActivityValues = activityData?.filter(
-    (activity) =>
-      activity.inventoryValue.gpcReferenceNumber === refNumberWithScope,
-  );
+  const { methodologies, directMeasure } = getMethodologies();
+
+  // extract the methodology used from the filtered scope
+
+  const [methodology, setMethodology] = useState<Methodology>();
+
+  const getfilteredActivityValues = useMemo(() => {
+    let methodologyId: string | null | undefined = null;
+    const filteredValues = activityData?.filter((activity) => {
+      let val =
+        activity.inventoryValue.gpcReferenceNumber === refNumberWithScope;
+      if (val && !methodologyId) {
+        methodologyId = activity.inventoryValue.inputMethodology;
+      }
+      return val;
+    });
+
+    if (methodologyId) {
+      let methodology =
+        methodologies.find((methodology) => methodology.id === methodologyId) ??
+        directMeasure;
+      setSelectedMethodology(methodologyId);
+      setIsMethodologySelected(true);
+      if (methodology && methodologyId)
+        setMethodology({
+          ...methodology,
+          fields: methodology.activities
+            ? methodology.activities
+            : (methodology as unknown as DirectMeasure)["extra-fields"],
+        });
+    }
+
+    return filteredValues;
+  }, [activityData, refNumberWithScope]);
+
+  function getMethodologies() {
+    const methodologies =
+      MANUAL_INPUT_HIERARCHY[refNumberWithScope]?.methodologies || [];
+    const directMeasure =
+      MANUAL_INPUT_HIERARCHY[refNumberWithScope]?.directMeasure;
+    return { methodologies, directMeasure };
+  }
 
   const inventoryValue = useMemo<InventoryValue | null>(() => {
     return (
@@ -127,15 +127,6 @@ const ActivityTab: FC<ActivityTabProps> = ({
 
   const activityValues =
     getActivityValuesByMethodology(getfilteredActivityValues) || [];
-  function getMethodologies() {
-    const methodologies =
-      MANUAL_INPUT_HIERARCHY[refNumberWithScope]?.methodologies || [];
-    const directMeasure =
-      MANUAL_INPUT_HIERARCHY[refNumberWithScope]?.directMeasure;
-    return { methodologies, directMeasure };
-  }
-
-  const { methodologies, directMeasure } = getMethodologies();
 
   const getSuggestedActivities = (): SuggestedActivity[] => {
     if (!selectedMethodology) return [];
@@ -160,32 +151,7 @@ const ActivityTab: FC<ActivityTabProps> = ({
   const changeMethodology = () => {
     setSelectedMethodology("");
     setIsMethodologySelected(false);
-    onChangeMethodologyClose();
   };
-
-  const {
-    isOpen: isAddActivityModalOpen,
-    onOpen: onAddActivityModalOpen,
-    onClose: onAddActivityModalClose,
-  } = useDisclosure();
-
-  const {
-    isOpen: isChangeMethodologyModalOpen,
-    onOpen: onChangeMethodologyOpen,
-    onClose: onChangeMethodologyClose,
-  } = useDisclosure();
-
-  const {
-    isOpen: isDeleteActivitiesModalOpen,
-    onOpen: onDeleteActivitiesModalOpen,
-    onClose: onDeleteActivitiesModalClose,
-  } = useDisclosure();
-
-  const {
-    isOpen: isDeleteActivityModalOpen,
-    onOpen: onDeleteActivityModalOpen,
-    onClose: onDeleteActivityModalClose,
-  } = useDisclosure();
 
   const suggestedActivities: SuggestedActivity[] = getSuggestedActivities();
 
@@ -193,46 +159,6 @@ const ActivityTab: FC<ActivityTabProps> = ({
     setIsChecked(!isUnavailableChecked);
   };
 
-  const handleActivityAdded = (
-    suggestedActivity: SuggestedActivity,
-    // ...args: any[]
-  ) => {
-    setSelectedActivity(suggestedActivity);
-
-    onAddActivityModalOpen();
-  };
-
-  const onDeleteActivity = (activity: ActivityValue) => {
-    setSelectedActivityValue(activity);
-    onDeleteActivityModalOpen();
-  };
-
-  const onEditActivity = (activity: ActivityValue) => {
-    setSelectedActivityValue(activity);
-    onAddActivityModalOpen();
-  };
-
-  const closeModals = () => {
-    setSelectedActivityValue(undefined);
-    onAddActivityModalClose();
-    onDeleteActivitiesModalClose();
-    onDeleteActivityModalClose();
-  };
-
-  function handleCardSelect(
-    disabled: boolean | undefined,
-    inputRequired: string[] | undefined,
-    id: string,
-    fields: any,
-  ) {
-    return () =>
-      handleMethodologySelected({
-        disabled: !!disabled,
-        inputRequired,
-        id,
-        fields,
-      });
-  }
   return (
     <>
       <TabPanel p="0" pt="48px">
@@ -253,429 +179,42 @@ const ActivityTab: FC<ActivityTabProps> = ({
             </Text>
           </Box>
         </Box>
-        {isMethodologySelected ? (
+        {isUnavailableChecked && <ScopeUnavailable t={t} />}
+        {!isUnavailableChecked && (
           <>
-            <Box
-              h="auto"
-              px="24px"
-              py="32px"
-              bg="base.light"
-              borderRadius="8px"
-            >
-              {" "}
-              {isMethodologySelected ? (
-                <Box>
-                  <Text
-                    fontFamily="heading"
-                    fontSize="10px"
-                    fontWeight="semibold"
-                    letterSpacing="widest"
-                    textTransform="uppercase"
-                    color="content.tertiary"
-                  >
-                    {t("methodology")}
-                  </Text>
-                  <Box display="flex" justifyContent="space-between">
-                    <Box>
-                      <HeadingText title={t(methodology?.id || "")} />
-                      <Text
-                        letterSpacing="wide"
-                        fontSize="body.lg"
-                        fontWeight="normal"
-                        color="interactive.control"
-                      >
-                        {t(methodology?.id + "-description")}
-                      </Text>
-                    </Box>
-                    <Box display="flex" alignItems="center">
-                      {(activityValues.length > 0 ||
-                        getInputMethodology(methodology?.id!)) !==
-                        "direct-measure" && (
-                        <Button
-                          data-testid="add-emission-data-button"
-                          onClick={onAddActivityModalOpen}
-                          title="Add Activity"
-                          leftIcon={<AddIcon h="16px" w="16px" />}
-                          h="48px"
-                          aria-label="activity-button"
-                          fontSize="button.md"
-                          gap="8px"
-                        >
-                          {t("add-emission-data")}
-                        </Button>
-                      )}
-                      <Popover>
-                        <PopoverTrigger>
-                          <IconButton
-                            icon={<MdMoreVert size="24px" />}
-                            aria-label="more-icon"
-                            variant="ghost"
-                            color="content.tertiary"
-                          />
-                        </PopoverTrigger>
-                        <PopoverContent
-                          w="auto"
-                          borderRadius="8px"
-                          shadow="2dp"
-                          px="0"
-                        >
-                          <PopoverArrow />
-                          <PopoverBody p="0px">
-                            <Box
-                              p="16px"
-                              display="flex"
-                              alignItems="center"
-                              gap="16px"
-                              _hover={{
-                                bg: "content.link",
-                                cursor: "pointer",
-                              }}
-                              className="group"
-                              onClick={onChangeMethodologyOpen}
-                            >
-                              <Icon
-                                className="group-hover:text-white"
-                                color="interactive.control"
-                                as={FaNetworkWired}
-                                h="24px"
-                                w="24px"
-                              />
-                              <Text
-                                className="group-hover:text-white"
-                                color="content.primary"
-                              >
-                                {t("change-methodology")}
-                              </Text>
-                            </Box>
-                            <Box
-                              p="16px"
-                              display="flex"
-                              alignItems="center"
-                              gap="16px"
-                              _hover={{
-                                bg: "content.link",
-                                cursor: "pointer",
-                              }}
-                              className="group"
-                              onClick={onDeleteActivitiesModalOpen}
-                            >
-                              <Icon
-                                className="group-hover:text-white"
-                                color="sentiment.negativeDefault"
-                                as={FiTrash2}
-                                h="24px"
-                                w="24px"
-                              />
-                              <Text
-                                className="group-hover:text-white"
-                                color="content.primary"
-                              >
-                                {t("delete-all-activities")}
-                              </Text>
-                            </Box>
-                          </PopoverBody>
-                        </PopoverContent>
-                      </Popover>
-                    </Box>
-                  </Box>
-                  <Box
-                    mt="48px"
-                    display="flex"
-                    flexDirection="column"
-                    gap="16px"
-                  >
-                    {activityValues?.length ? (
-                      <Box>
-                        {getInputMethodology(methodology?.id!) ===
-                        "direct-measure" ? (
-                          <DirectMeasureTable
-                            t={t}
-                            referenceNumber={refNumberWithScope}
-                            activityData={activityValues}
-                            onDeleteActivity={onDeleteActivity}
-                            onEditActivity={onEditActivity}
-                          />
-                        ) : (
-                          <ActivityAccordion
-                            t={t}
-                            referenceNumber={refNumberWithScope}
-                            activityData={activityValues}
-                            showActivityModal={onAddActivityModalOpen}
-                            methodologyId={methodology?.id}
-                            onDeleteActivity={onDeleteActivity}
-                            onEditActivity={onEditActivity}
-                          />
-                        )}
-                        <Box
-                          w="full"
-                          borderTopWidth="3px"
-                          borderColor="interactive.secondary"
-                          py="32px"
-                          px="48px"
-                        >
-                          <Box display="flex" justifyContent="space-between">
-                            <Text
-                              fontFamily="heading"
-                              fontWeight="semibold"
-                              fontSize="headline.md"
-                            >
-                              {t("total-emissions")}
-                            </Text>
-                            <Text
-                              fontFamily="heading"
-                              fontWeight="semibold"
-                              fontSize="headline.md"
-                            >
-                              {convertKgToTonnes(totalEmissions)}
-                            </Text>
-                          </Box>
-                        </Box>
-                      </Box>
-                    ) : (
-                      <>
-                        {suggestedActivities.length ? (
-                          <>
-                            <Text
-                              fontFamily="heading"
-                              fontSize="title.md"
-                              fontWeight="semibold"
-                              color="content.secondary"
-                            >
-                              {t("activity-suggestion")}
-                            </Text>
-                            <Box className="flex flex-col gap-4">
-                              {suggestedActivities.map((suggestedActivity) => {
-                                const { id } = suggestedActivity;
-                                const prefillKey =
-                                  suggestedActivity.prefills[0]?.key;
-                                const prefillValue =
-                                  suggestedActivity.prefills[0]?.value;
-                                return (
-                                  <SuggestedActivityCard
-                                    key={id}
-                                    id={id}
-                                    prefillKey={prefillKey}
-                                    prefillValue={prefillValue}
-                                    t={t}
-                                    isSelected={selectedActivity?.id === id}
-                                    onActivityAdded={(...args) =>
-                                      handleActivityAdded(
-                                        suggestedActivity,
-                                        // ...args,
-                                      )
-                                    }
-                                  />
-                                );
-                              })}
-                            </Box>
-                          </>
-                        ) : (
-                          <Card
-                            w="full"
-                            bg="background.backgroundLight"
-                            shadow="none"
-                            h="100px"
-                            flexDir="row"
-                            p="24px"
-                            justifyContent="space-between"
-                          >
-                            <Box display="flex" gap="12px">
-                              <Box>
-                                <DataConnectIcon />
-                              </Box>
-                              <Box display="flex" flexDir="column" gap="8px">
-                                <Text
-                                  color="interactive.secondary"
-                                  fontSize="title.md"
-                                  fontWeight="bold"
-                                  lineHeight="24px"
-                                  fontFamily="heading"
-                                >
-                                  {t("add-emissions-data-title")}
-                                </Text>
-                                <Text
-                                  color="content.tertiary"
-                                  fontSize="body.md"
-                                  fontWeight="semibold"
-                                  lineHeight="20px"
-                                  letterSpacing="wide"
-                                >
-                                  {t("add-emissions-data-subtext")}
-                                </Text>
-                              </Box>
-                            </Box>
-                            <Box>
-                              <Button
-                                onClick={onAddActivityModalOpen}
-                                data-testid="add-emission-data-button"
-                                title={t("add-emission-data")}
-                                leftIcon={<AddIcon h="16px" w="16px" />}
-                                h="48px"
-                                aria-label="activity-button"
-                                fontSize="button.md"
-                                gap="8px"
-                              >
-                                {t("add-emission-data-btn")}
-                              </Button>
-                            </Box>
-                          </Card>
-                        )}
-                      </>
-                    )}
-                  </Box>
-                </Box>
-              ) : (
-                <Box
-                  display="flex"
-                  alignItems="center"
-                  justifyContent="space-between"
-                  mb="8px"
-                >
-                  <HeadingText title={t("select-methodology-title")} />
-                </Box>
-              )}
-            </Box>
-          </>
-        ) : (
-          <Box>
-            {isUnavailableChecked ? (
-              <ScopeUnavailable t={t} />
-            ) : (
-              <Box>
-                {isUnavailableChecked && (
-                  <Box>
-                    <HeadingText title={t("scope-unavailable")} />
-                    <Text
-                      letterSpacing="wide"
-                      fontSize="body.lg"
-                      fontWeight="normal"
-                      color="interactive.control"
-                      mt="8px"
-                    >
-                      {t("scope-unavailable-description")}
-                    </Text>
-                    <Box mt="48px">
-                      {!hasActivityData && (
-                        <HeadingText title={t("select-methodology-title")} />
-                      )}
-                    </Box>
-                  </Box>
-                )}
-                {isMethodologySelected ? (
-                  ""
-                ) : (
-                  <Box>
-                    <Text
-                      letterSpacing="wide"
-                      fontSize="body.lg"
-                      fontWeight="normal"
-                      color="interactive.control"
-                    >
-                      <Trans t={t} i18nKey="add-data-manually-desciption">
-                        To add your inventory data manually, select the
-                        methodology used to collect the data and calculate your
-                        emissions.{" "}
-                        <Link
-                          href="https://ghgprotocol.org/ghg-protocol-cities"
-                          color="content.link"
-                          fontWeight="bold"
-                          textDecoration="underline"
-                          target="_blank"
-                          rel="noreferrer noopener"
-                        >
-                          Learn more
-                        </Link>{" "}
-                        about methodologies
-                      </Trans>
-                    </Text>
-                    <Text
-                      fontWeight="bold"
-                      fontSize="title.md"
-                      fontFamily="heading"
-                      pt="48px"
-                      pb="24px"
-                    >
-                      {t("select-methodology")}
-                    </Text>
-                    <SimpleGrid minChildWidth="250px" spacing={4}>
-                      {(methodologies || []).map(
-                        ({ id, disabled, activities, inputRequired }) => (
-                          <MethodologyCard
-                            id={id}
-                            key={id}
-                            inputRequired={inputRequired}
-                            isSelected={selectedMethodology === id}
-                            disabled={!!disabled}
-                            t={t}
-                            handleCardSelect={handleCardSelect(
-                              disabled,
-                              inputRequired,
-                              id,
-                              activities,
-                            )}
-                          />
-                        ),
-                      )}
-                      {directMeasure?.id ? (
-                        <MethodologyCard
-                          id={directMeasure.id}
-                          key={directMeasure.id}
-                          isSelected={selectedMethodology === directMeasure.id}
-                          t={t}
-                          handleCardSelect={handleCardSelect(
-                            false,
-                            ["emissions-data"],
-                            directMeasure.id,
-                            directMeasure["extra-fields"],
-                          )}
-                          disabled={false}
-                        />
-                      ) : null}
-                    </SimpleGrid>
-                  </Box>
-                )}
+            {isMethodologySelected ? (
+              <Box
+                h="auto"
+                px="24px"
+                py="32px"
+                bg="base.light"
+                borderRadius="8px"
+              >
+                {" "}
+                <EmissionDataSection
+                  t={t}
+                  methodology={methodology}
+                  inventoryId={inventoryId}
+                  subsectorId={subsectorId}
+                  refNumberWithScope={refNumberWithScope}
+                  activityValues={activityValues}
+                  suggestedActivities={suggestedActivities}
+                  totalEmissions={totalEmissions}
+                  changeMethodology={changeMethodology}
+                  inventoryValue={inventoryValue}
+                />
               </Box>
+            ) : (
+              <SelectMethodology
+                t={t}
+                methodologies={methodologies}
+                handleMethodologySelected={handleMethodologySelected}
+                directMeasure={directMeasure}
+              />
             )}
-          </Box>
+          </>
         )}
       </TabPanel>
-      <ActivityFormModal
-        t={t}
-        isOpen={isAddActivityModalOpen}
-        onClose={onAddActivityModalClose}
-        hasActivityData={hasActivityData}
-        setHasActivityData={setHasActivityData}
-        methodology={methodology!}
-        inventoryId={inventoryId}
-        inventoryValue={inventoryValue}
-        selectedActivity={selectedActivity}
-        referenceNumber={refNumberWithScope}
-        edit={!!selectedActivityValue}
-        targetActivityValue={selectedActivityValue as ActivityValue}
-        resetSelectedActivityValue={() => setSelectedActivityValue(undefined)}
-      />
-
-      <ChangeMethodology
-        t={t}
-        onClose={onChangeMethodologyClose}
-        isOpen={isChangeMethodologyModalOpen}
-        onChangeClicked={changeMethodology}
-      />
-      <DeleteAllActivitiesModal
-        t={t}
-        isOpen={isDeleteActivitiesModalOpen}
-        onClose={onDeleteActivitiesModalClose}
-        inventoryId={inventoryId}
-        subsectorId={subsectorId}
-      />
-      <DeleteActivityModal
-        t={t}
-        isOpen={isDeleteActivityModalOpen}
-        onClose={onDeleteActivityModalClose}
-        selectedActivityValue={selectedActivityValue as ActivityValue}
-        inventoryId={inventoryId}
-        resetSelectedActivityValue={() => setSelectedActivityValue(undefined)}
-      />
     </>
   );
 };
