+++ app/src/i18n/locales/en/data.json
@@ -783,5 +783,7 @@
   "wastewater-inside-generated": "Quantity of wastewater generated inside the city",
   "total-organic-sludge-removed": "Total organic sludge removed",
   "population-or-industrial-production": "Population or industrial production",
-  "wastewater-outside-generated": "Quantity of wastewater generated outside the city"
+  "wastewater-outside-generated": "Quantity of wastewater generated outside the city",
+  "change-methodology-success": "Methodology successfully changed",
+  "change-methodology-error": "An issue occured with changing the methodology"
 }
