+++ app/src/services/api.ts
@@ -1,25 +1,25 @@
 import {
-  type UserAttributes,
   type CityAttributes,
   type InventoryAttributes,
   type InventoryValueAttributes,
   PopulationAttributes,
+  type UserAttributes,
 } from "@/models/init-models";
 import type { BoundingBox } from "@/util/geojson";
 import type {
   ConnectDataSourceQuery,
   ConnectDataSourceResponse,
   DataSourceResponse,
+  EmissionsFactorResponse,
   InventoryProgressResponse,
   InventoryResponse,
-  InventoryValueUpdateQuery,
   InventoryValueResponse,
+  InventoryValueUpdateQuery,
   InventoryWithCity,
-  UserInfoResponse,
-  UserFileResponse,
-  EmissionsFactorResponse,
-  UserInviteResponse,
   RequiredScopesResponse,
+  UserFileResponse,
+  UserInfoResponse,
+  UserInviteResponse,
 } from "@/util/types";
 import type { GeoJSON } from "geojson";
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
@@ -502,11 +502,16 @@ export const api = createApi({
       invalidatesTags: ["ActivityValue"],
     }),
     deleteAllActivityValues: builder.mutation({
-      query: (data: { inventoryId: string; subSectorId: string }) => ({
+      query: (data: {
+        inventoryId: string;
+        subSectorId: string;
+        gpcReferenceNumber: string;
+      }) => ({
         method: "DELETE",
         url: `/inventory/${data.inventoryId}/activity-value`,
         params: {
           subSectorId: data.subSectorId,
+          gpcReferenceNumber: data.gpcReferenceNumber,
         },
       }),
       transformResponse: (response: any) => response.data,
